+++ src/Sandertv/BlockSniper/brush/shapes/SphereShape.php
@@ -13,7 +13,7 @@ use Sandertv\BlockSniper\Loader;
 
 class SphereShape extends BaseShape {
 	
-	public function __construct(Loader $main, Player $player, Level $level, float $radius = null, Position $center = null, bool $hollow = false) {
+	public function __construct(Loader $main, Player $player, Level $level, int $radius = null, Position $center = null, bool $hollow = false) {
 		parent::__construct($main);
 		$this->level = $level;
 		$this->radius = $radius;
